modified code of apply_patchfile.py and download_patchfile.py

+++ vim74/code/download_patchfile.py
@@ -5,10 +5,13 @@
 import os
 
 min_patch_num = 1
-max_patch_num = 450
+max_patch_num = 550
 
 for i in range(min_patch_num, max_patch_num):
     patch_file = "7.4.{0:003}".format(i)
+
+    if os.path.exists(patch_file):
+        continue
 
     print("Start download patch file: " + patch_file)
     download_cmd = 'curl -o {0} ftp://ftp.vim.org/pub/vim/patches/7.4/{1}'.format(patch_file, patch_file)
